Move ActiveContractsInMemory to a separate file
